requirements/django.txt
django==3.1.4
djangorestframework==3.12.2
python-memcached==1.59
djangorestframework-jwt==1.11.0
psycopg2==2.8.5
uWSGI==2.0.19
